Add files via upload
The starting point was data on a company that sells coffee in different presentations, of different types, to clients in 3 countries and that has a loyalty program with data from 2019 to 2022.  The aim is to be able to know the specific sales by periods, being able to use filters in relation to the type of coffee, the size if the clients belong or not to the loyalty program and to see the percentage that each one represents according to the country.